requirements.txt
langchain
numpy
pandas
ipykernel
python-dotenv
langchain_openai
langchain_core
langchain_community
pypdf
langchain_text_splitters
faiss-cpu
streamlit
googletrans
datasets
torch
librosa
accelerate>=0.26.0
pydub
audioop-lts